experiment with variance

+++ nncf/quantization/algorithms/weight_compression/mixed_precision.py
@@ -10,6 +10,7 @@
 # limitations under the License.
 
 from abc import abstractmethod
+from pathlib import Path
 from typing import Dict, List, Optional, TypeVar
 
 import pandas as pd
@@ -77,10 +78,16 @@ class MixedPrecisionCriterion:
         """
         Assigns quantization precision based on computed layers' sensitivities, ratio of parameters.
         """
-        scores = self._calc_sensitivity()
-        df = pd.DataFrame(scores)
-        df.to_csv('scores.csv')
-        # assert False
+        # NOISE_FILE = Path('max_var_scores.csv')
+        NOISE_FILE = Path('/home/nlyaly/projects/lm-evaluation-harness/scores.csv')
+        if NOISE_FILE.exists():
+            df = pd.read_csv(NOISE_FILE)
+            scores = list(df[df.columns[1]])
+        else:
+            scores = self._calc_sensitivity()
+            df = pd.DataFrame(scores)
+            df.to_csv('scores.csv')
+            # assert False
         num_all_weights = sum(wp.num_weights for wp in self._weight_params)
 
         indexes_of_layers_in_ascending_order_of_scores = [
